FieldSpec and ImportSpec are interfaces with default implementations
