+++ extensions/jest/src/test/java/io/doov/ts/ast/test/TypeScriptAnyMatchTest.java
@@ -23,6 +23,7 @@ import static io.doov.ts.ast.test.JestExtension.parseAs;
 import static io.doov.ts.ast.test.TypeScriptAnyMatchTest.EnumTest.VAL1;
 import static io.doov.ts.ast.test.TypeScriptAnyMatchTest.EnumTest.VAL2;
 import static io.doov.ts.ast.test.TypeScriptAnyMatchTest.EnumTest.VAL3;
+import static io.doov.ts.ast.writer.DefaultImportSpec.newImport;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
@@ -37,7 +38,6 @@ import io.doov.core.dsl.field.types.EnumFieldInfo;
 import io.doov.core.dsl.lang.Result;
 import io.doov.core.dsl.lang.StepCondition;
 import io.doov.core.dsl.runtime.GenericModel;
-import io.doov.ts.ast.writer.ImportSpec;
 import io.doov.tsparser.TypeScriptParser;
 
 class TypeScriptAnyMatchTest {
@@ -167,7 +167,7 @@ class TypeScriptAnyMatchTest {
 
     @AfterAll
     static void tearDown() {
-        jestExtension.getJestTestSpec().getImports().add(new ImportSpec("BooleanFunction", "doov"));
+        jestExtension.getJestTestSpec().getImports().add(newImport("doov", "BooleanFunction"));
         jestExtension.getJestTestSpec().getTestStates().add("const alwaysFalseA = DOOV.lift(BooleanFunction, false);");
         jestExtension.getJestTestSpec().getTestStates().add("const alwaysTrueA = DOOV.lift(BooleanFunction, true);");
         jestExtension.getJestTestSpec().getTestStates().add("enum EnumTest { VAL1, VAL2, VAL3 }");
